Sorry, I'm not good at only making one or two commits.  I keep forgetting something.
Anyway, I installed my edited menu as firmware.  It only sacrificed 1024 bytes of flash memory.
Second off, I included a patch in GraphicStartup that prevents it from crashing when it tries to make a file with more than 16 characters in it's name the default.  I'm not going to work on this much for a couple days in order to catch up on homework.  But feel free to comment on, suggest new ideas, or improve what I have.  Thxs

Abram Early

git-svn-id: http://svn.code.sf.net/p/lejos/code/trunk@4163 06c8f799-c779-4783-8af7-179afedb67c8