Finalized the wrapping.  So heres a summary of what it does now. 1 Item = No animation, No Wrapping 2 Items = No Wrapping 3 Items = Shows only the three icons once. Wraps 4 Items = Shows only the four icons once. (The four icon takes up the far right space in the menu) Wraps 5 Items+ = Normal Wrapping
Other features included,
Graphic Yes/No Menu

There is a HUGE error in the Startup Menu.  It can't delete a file correctly.  It runs into a weird error when a file is deleted.  (When there is more than one.)  And then, all the others files can't run anymore because they get an error on startup.  I think it's deleting more than it should.  I'm not going to try and fix that.  Somebody else should do that soon.